👕 refactor global mixin to utils module
