+++ src/modals/EmailChange.vue
@@ -62,6 +62,7 @@
 <script>
 import { defineComponent } from 'vue';
 import firebase from 'firebase/compat/app';
+import { throwError } from '@/utils.js';
 
 export default defineComponent({
 	name: 'email-change',
@@ -118,9 +119,9 @@ export default defineComponent({
 									text: this.$t('toast.verficationSentText'),
 									type: 'primary'
 								});
-							}).catch((error) => this.throwError(error));
-						}).catch((error) => this.throwError(error));
-					}).catch((error) => this.throwError(error));
+							}).catch((error) => throwError(error));
+						}).catch((error) => throwError(error));
+					}).catch((error) => throwError(error));
 				}).catch(() => {
 					this.error.currentpassword.wrong = true;
 					this.$notify({
