+++ src/modals/SetlistDelete.vue
@@ -24,6 +24,7 @@
 
 <script>
 import { defineComponent } from 'vue';
+import { throwError } from '@/utils.js';
 
 export default defineComponent({
 	name: 'setlist-delete',
@@ -45,7 +46,7 @@ export default defineComponent({
 					text: this.$t('toast.setlistDeletedText'),
 					type: 'primary'
 				});
-			}).catch((error) => this.throwError(error));
+			}).catch((error) => throwError(error));
 		},
 	}
 });
