+++ src/modals/SongDelete.vue
@@ -24,6 +24,7 @@
 
 <script>
 import { defineComponent } from 'vue';
+import { throwError } from '@/utils.js';
 
 export default defineComponent({
 	name: 'song-delete',
@@ -59,7 +60,7 @@ export default defineComponent({
 					text: this.$t('toast.songDeletedText'),
 					type: 'primary'
 				});
-			}).catch((error) => this.throwError(error));
+			}).catch((error) => throwError(error));
 		},
 	}
 });
