+++ src/views/Profile.vue
@@ -109,6 +109,7 @@
 import { computed } from 'vue';
 import { useI18n } from "vue-i18n";
 const { t } = useI18n();
+import { initials } from '@/utils.js';
 
 // get icons
 import {
